backend/requirements.txt
# Django 5.2 LTS (soportado hasta abril 2028).
# 5.0 llegó a EOL en abril 2025.
Django==5.2.12
djangorestframework==3.15.2
# psycopg2 compilado (recomendado para producción; psycopg2-binary no es estable en prod).
# Requiere libpq-dev en el sistema (ya presente en el Dockerfile).
psycopg2==2.9.11
django-cors-headers==4.6.0
celery==5.4.0
# sentry-sdk v2: API estable, soporte activo. Requiere traces_sample_rate en init.
sentry-sdk==2.19.2
# En contenedores no es seguro depender de schedulers con archivo local:
# el estado se pierde al reiniciar; django-celery-beat persiste en PostgreSQL.
django-celery-beat==2.9.0
redis==5.2.1
python-decouple==3.8
# Pillow: corrige vulnerabilidades de seguridad
Pillow==12.1.1
pgvector==0.3.6
# requests 2.32.x: corrige CVE-2024-35195 y CVE-2024-47081
requests==2.32.5

# JWT para autenticación (corrige CVE-2024-22513)
djangorestframework-simplejwt==5.5.1

# ONNX Runtime para embeddings locales
onnxruntime==1.20.1
numpy==1.26.4

# Parser de fechas robusto
python-dateutil==2.9.0

# RSS feed parsing (Goodreads sync)
feedparser==6.0.12
beautifulsoup4==4.12.3

# ML para clustering y moderación
scikit-learn==1.5.2

# Multilingual ONNX tokenization
tokenizers==0.21.0

# WSGI server (pinned — antes se instalaba suelto en Dockerfile)
gunicorn==23.0.0

# Test coverage reporting
coverage==7.6.10
